lane-dp: extract lane execution into TTLaneStepExecutor
model_runner.py owned both generic TT execution and the lane-DP slot
geometry, input building, and output extraction. Move the lane-specific
path into lane_execution.py (with TTModelInput/sampling in model_input.py
and logprob packing in logprobs.py) so the runner only delegates. Guard
lane-DP against request-specific RoPE: it assumes front-packed rows while
lane slots are sparse. Add host tests for the sparse-slot index logic.
